// src/utils/types.ts
export interface JsonData {
    id: number;
    first: string;
    last: string;
    dob: string;
    gender: string;
    email: string;
    picture: string;
    country: string;
    description: string;
}

export interface Item {
    id: number;
    first:string;
    gender: string;
    country: string;
    description: string;
  }
  